manage preload utils
