+++ config/webpack/webpack.common.js
@@ -44,9 +44,9 @@ commonConfig = {
       ".css"
     ],
     alias: {
-      // "react": "preact/compat",
-      // "react-dom/test-utils": "preact/test-utils",
-      // "react-dom": "preact/compat",
+      react: "preact/compat",
+      "react-dom/test-utils": "preact/test-utils",
+      "react-dom": "preact/compat"
     },
     modules: [paths.nodeModules, paths.src]
   },
