--- src/lib/solidify/style/reset.less
@@ -1,19 +1,0 @@
-/**
- * Hard reset
- */
-* {
-  margin: 0;
-  padding: 0;
-}
-
-/**
- * Default Box sizing
- */
-html {
-  box-sizing: border-box;
-}
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
-}
